Add FRIBDAQ readout.
*  Added parser_callbacks header to define the callbacks we'll be
   implementing.
*  Remove the definition of extra state to the parser_callback.h
*  Add an include for that header int the main file.
